Renaming package "Whitelist" to "whitelist", adding corresponding package in test project
