+++ rskj-core/src/main/java/co/rsk/peg/BridgeSupport.java
@@ -31,10 +31,10 @@ import co.rsk.config.RskSystemProperties;
 import co.rsk.core.RskAddress;
 import co.rsk.crypto.Keccak256;
 import co.rsk.panic.PanicProcessor;
-import co.rsk.peg.Whitelist.LockWhitelist;
-import co.rsk.peg.Whitelist.LockWhitelistEntry;
-import co.rsk.peg.Whitelist.OneOffWhiteListEntry;
-import co.rsk.peg.Whitelist.UnlimitedWhiteListEntry;
+import co.rsk.peg.whitelist.LockWhitelist;
+import co.rsk.peg.whitelist.LockWhitelistEntry;
+import co.rsk.peg.whitelist.OneOffWhiteListEntry;
+import co.rsk.peg.whitelist.UnlimitedWhiteListEntry;
 import co.rsk.peg.utils.BridgeEventLogger;
 import co.rsk.peg.utils.BtcTransactionFormatUtils;
 import co.rsk.peg.utils.PartialMerkleTreeFormatUtils;
