+++ rskj-core/src/test/java/co/rsk/peg/BridgeSupportTest.java
@@ -34,8 +34,8 @@ import co.rsk.core.BlockDifficulty;
 import co.rsk.core.RskAddress;
 import co.rsk.crypto.Keccak256;
 import co.rsk.db.RepositoryImpl;
-import co.rsk.peg.Whitelist.LockWhitelist;
-import co.rsk.peg.Whitelist.OneOffWhiteListEntry;
+import co.rsk.peg.whitelist.LockWhitelist;
+import co.rsk.peg.whitelist.OneOffWhiteListEntry;
 import co.rsk.peg.simples.SimpleBlockChain;
 import co.rsk.peg.simples.SimpleRskTransaction;
 import co.rsk.peg.simples.SimpleWallet;
