+++ rskj-core/src/test/java/co/rsk/peg/performance/LockWhitelistTest.java
@@ -24,9 +24,9 @@ import co.rsk.bitcoinj.store.BtcBlockStore;
 import co.rsk.config.TestSystemProperties;
 import co.rsk.peg.Bridge;
 import co.rsk.peg.BridgeStorageProvider;
-import co.rsk.peg.Whitelist.LockWhitelist;
+import co.rsk.peg.whitelist.LockWhitelist;
 import co.rsk.peg.RepositoryBlockStore;
-import co.rsk.peg.Whitelist.OneOffWhiteListEntry;
+import co.rsk.peg.whitelist.OneOffWhiteListEntry;
 import org.ethereum.core.Repository;
 import org.ethereum.crypto.ECKey;
 import org.ethereum.vm.PrecompiledContracts;
